Improve Toggle implementation and separate variants.
