+++ src/core/Form/Form.tsx
@@ -1,4 +1,9 @@
-export { ToggleProps, ToggleInput, ToggleButton } from './Toggle';
+export {
+  ToggleInput,
+  ToggleInputProps,
+  ToggleButton,
+  ToggleButtonProps,
+} from './Toggle';
 export { TextInput, TextInputProps } from './TextInput/TextInput';
 export { SearchInput, SearchInputProps } from './SearchInput/SearchInput';
 export { Checkbox, CheckboxProps } from './Checkbox/Checkbox';
