+++ src/core/Form/Toggle/ToggleInput/ToggleInput.md
@@ -8,21 +8,21 @@ const [checked, setChecked] = useState(false);
 <>
   <ToggleInput
     defaultChecked
-    onClick={({ toggleState }) => console.log(toggleState)}
+    onChange={({ toggleState }) => console.log(toggleState)}
   >
     Checked enabled using input
   </ToggleInput>
 
   <ToggleInput
     disabled
-    onClick={({ toggleState }) => console.log(toggleState)}
+    onChange={({ toggleState }) => console.log(toggleState)}
   >
     Unchecked disabled using input
   </ToggleInput>
 
   <ToggleInput
     checked={checked}
-    onClick={({ toggleState }) => {
+    onChange={({ toggleState }) => {
       if (window.confirm('Change Toggle state?')) {
         setChecked(toggleState);
       }
