+++ src/index.tsx
@@ -14,9 +14,10 @@ export {
   CheckboxProps,
   TextInput,
   TextInputProps,
-  ToggleProps,
   ToggleInput,
+  ToggleInputProps,
   ToggleButton,
+  ToggleButtonProps,
   SearchInput,
   SearchInputProps,
   RadioButton,
